refactor: GameRule 패키지 game 에서 employee로 이동
