+++ src/main/kotlin/racinggame/domain/employee/GameGuide.kt
@@ -1,6 +1,5 @@
 package racinggame.domain.employee
 
-import racinggame.domain.game.GameRule
 import racinggame.domain.player.User
 
 data class GameGuide(
